Separated l0 and l1
