+++ imap_processing/tests/idex/test_decom.py
@@ -1,17 +1,15 @@
 """Tests the decommutation process for IDEX CCSDS Packets."""
 
-from pathlib import Path
-
 import numpy as np
 import pytest
 
 from imap_processing import imap_module_directory
-from imap_processing.idex.idex_packet_parser import PacketParser
+from imap_processing.idex.l1.idex_l1 import PacketParser
 
 
 @pytest.fixture(scope="session")
 def decom_test_data():
-    test_file = Path(
+    test_file = (
         f"{imap_module_directory}/tests/idex/imap_idex_l0_raw_20230725_v001.pkts"
     )
     return PacketParser(test_file, "001")
